Bug 502076 - Fix test failures in CompletionTests2.testBug6930_23
The fingerprint tests weren't taking into account the resolution
of the filesystem timestamps, so files were getting incorrectly marked
as unchanged if they changed more rapidly than the filesystem could
record. Fixed by adding a new sentinel to FileFingerprint indicating
that the timestamp should be ignored and the hashcode should always
be used in circumstances where the timestamp is too close to the current
time to reliably identify future changes.

Change-Id: I5b0f964016c3296ce6c10eb2202be07178350c7a
Signed-off-by: Stefan Xenos <sxenos@gmail.com>